// src/interfaces/meragverse.enums.ts
export enum AstralObjects {
  SPACE = "SPACE",
  POLYANET = "POLYANET",
  BLUE_SOLOON = "BLUE_SOLOON",
  RED_SOLOON = "RED_SOLOON",
  PURPLE_SOLOON = "PURPLE_SOLOON",
  WHITE_SOLOON = "WHITE_SOLOON",
  UP_COMETH = "UP_COMETH",
  DOWN_COMETH = "DOWN_COMETH",
  LEFT_COMETH = "LEFT_COMETH",
  RIGHT_COMETH = "RIGHT_COMETH",
}